Revert usage of glimmer/build for building (for now)
glimmer/build is still somewhat beta, and it doesn't provide a bundled
AMD or globals build, which we still use in some places (in ember-mobiledoc-editor,
e.g.)
