+++ .gitignore
@@ -2,4 +2,3 @@ dist/
 tmp/
 bower_components/
 node_modules/
-tests/
