+++ testem.json
@@ -1,8 +1,8 @@
 {
   "framework": "qunit",
-  "test_page": "tests/index.html",
-  "before_tests": "npm run build:tests",
-  "src_files": ["src/**/*", "test/**/*"],
+  "test_page": "dist/tests/index.html",
+  "before_tests": "npm run build",
+  "src_files": ["src/**/*.js", "test/**/*.js"],
   "launch_in_ci": ["PhantomJS"],
   "launch_in_dev": ["PhantomJS", "Chrome"]
 }
