+++ tests/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>mobiledoc dom renderer tests</title>
+
+  @@QUNIT_CSS
+</head>
+<body>
+
+  @@QUNIT_DOM
+
+  @@QUNIT_JS
+
+  <script src="/testem.js"></script>
+
+  @@LOADER_JS
+
+  <script src="../amd/mobiledoc-dom-renderer.js"></script>
+
+  @@AMD_TEST_JS
+  <script src="./simple-dom.js"></script>
+
+  @@TEST_LOADER_JS
+</body>
+</html>
