+++ tests/unit/renderers/0-3-test.js
@@ -1,24 +1,30 @@
-/* global QUnit */
+/* global QUnit, SimpleDOM */
 
-import Renderer from '../../../lib/index';
-import { RENDER_TYPE } from '../../../lib/index';
+import Renderer from 'mobiledoc-dom-renderer';
+import { RENDER_TYPE } from 'mobiledoc-dom-renderer';
+import ImageCard from 'mobiledoc-dom-renderer/cards/image';
 import {
   MARKUP_SECTION_TYPE,
   LIST_SECTION_TYPE,
   CARD_SECTION_TYPE,
   IMAGE_SECTION_TYPE
-} from '../../../lib/utils/section-types';
+} from 'mobiledoc-dom-renderer/utils/section-types';
 import {
   innerHTML,
   outerHTML,
   childNodesLength,
   escapeQuotes
 } from '../../helpers/dom';
-import SimpleDOM from 'simple-dom';
+import {
+  MARKUP_MARKER_TYPE,
+  ATOM_MARKER_TYPE
+} from 'mobiledoc-dom-renderer/utils/marker-types';
 
 const { test, module } = QUnit;
-const MOBILEDOC_VERSION = '0.2.0';
+const MOBILEDOC_VERSION_0_3_0 = '0.3.0';
+const MOBILEDOC_VERSION_0_3_1 = '0.3.1';
 const dataUri = "data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACwAAAAAAQABAAACAkQBADs=";
+
 
 let renderer;
 
@@ -28,27 +34,27 @@ function generateTests() {
 
 test('renders an empty mobiledoc', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
-    sections: [
-      [], // markers
-      []  // sections
-    ]
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
+    sections: []
   };
   let { result: rendered } = renderer.render(mobiledoc);
 
   assert.ok(!!rendered, 'renders result');
-  assert.ok(!rendered.firstChild, 'has no sections');
+  assert.equal(childNodesLength(rendered), 0, 'has no sections');
 });
 
 test('renders a mobiledoc without markups', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [], // markers
-      [   // sections
-        [MARKUP_SECTION_TYPE, 'P', [
-          [[], 0, 'hello world']]
-        ]
+      [MARKUP_SECTION_TYPE, 'P', [
+        [MARKUP_MARKER_TYPE, [], 0, 'hello world']]
       ]
     ]
   };
@@ -62,15 +68,15 @@ test('renders a mobiledoc without markups', (assert) => {
                'renders the text');
 });
 
-test('renders markup section "pull-quote" as div with class', (assert) => {
+test('renders 0.3.0 markup section "pull-quote" as div with class', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [], // markers
-      [   // sections
-        [MARKUP_SECTION_TYPE, 'pull-quote', [
-          [[], 0, 'hello world']]
-        ]
+      [MARKUP_SECTION_TYPE, 'pull-quote', [
+        [MARKUP_MARKER_TYPE, [], 0, 'hello world']]
       ]
     ]
   };
@@ -82,18 +88,15 @@ test('renders markup section "pull-quote" as div with class', (assert) => {
   assert.equal(outerHTML(sectionEl), '<div class="pull-quote">hello world</div>');
 });
 
-
-test('renders a mobiledoc with simple (no attributes) markup', (assert) => {
+test('renders 0.3.1 markup section "pull-quote" as div with class', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_1,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [        // markers
-        ['B'],
-      ],
-      [        // sections
-        [MARKUP_SECTION_TYPE, 'P', [
-          [[0], 1, 'hello world']]
-        ]
+      [MARKUP_SECTION_TYPE, 'pull-quote', [
+        [MARKUP_MARKER_TYPE, [], 0, 'hello world']]
       ]
     ]
   };
@@ -102,22 +105,63 @@ test('renders a mobiledoc with simple (no attributes) markup', (assert) => {
                'renders 1 section');
   let sectionEl = rendered.firstChild;
 
-  assert.equal(sectionEl.firstChild.tagName, 'B');
-  assert.equal(sectionEl.firstChild.firstChild.nodeValue, 'hello world');
+  assert.equal(outerHTML(sectionEl), '<div class="pull-quote">hello world</div>');
+});
+
+test('renders markup section "aside"', (assert) => {
+  let mobiledoc = {
+    version: MOBILEDOC_VERSION_0_3_1,
+    atoms: [],
+    cards: [],
+    markups: [],
+    sections: [
+      [MARKUP_SECTION_TYPE, 'aside', [
+        [MARKUP_MARKER_TYPE, [], 0, 'hello world']]
+      ]
+    ]
+  };
+  let { result: rendered } = renderer.render(mobiledoc);
+  assert.equal(childNodesLength(rendered), 1,
+               'renders 1 section');
+  let sectionEl = rendered.firstChild;
+
+  assert.equal(outerHTML(sectionEl), '<aside>hello world</aside>');
+});
+
+test('renders a mobiledoc with simple (no attributes) markup', (assert) => {
+  let mobiledoc = {
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [
+      ['B']
+    ],
+    sections: [
+      [MARKUP_SECTION_TYPE, 'P', [
+        [MARKUP_MARKER_TYPE, [0], 1, 'hello world']]
+      ]
+    ]
+  };
+  let { result: rendered } = renderer.render(mobiledoc);
+  assert.equal(childNodesLength(rendered), 1,
+               'renders 1 section');
+  let sectionEl = rendered.firstChild;
+
+  assert.equal(innerHTML(sectionEl), '<b>hello world</b>');
 });
 
 test('renders a mobiledoc with complex (has attributes) markup', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [
+      ['A', ['href', 'http://google.com']],
+    ],
     sections: [
-      [        // markers
-        ['A', ['href', 'http://google.com']],
-      ],
-      [        // sections
-        [MARKUP_SECTION_TYPE, 'P', [
-            [[0], 1, 'hello world']
-        ]]
-      ]
+      [MARKUP_SECTION_TYPE, 'P', [
+        [MARKUP_MARKER_TYPE, [0], 1, 'hello world']
+      ]]
     ]
   };
   let { result: rendered } = renderer.render(mobiledoc);
@@ -125,73 +169,72 @@ test('renders a mobiledoc with complex (has attributes) markup', (assert) => {
                'renders 1 section');
   let sectionEl = rendered.firstChild;
 
-  assert.equal(innerHTML(sectionEl), '<a href="http://google.com">hello world</a>');
   assert.equal(innerHTML(sectionEl), '<a href="http://google.com">hello world</a>');
 });
 
 test('renders a mobiledoc with multiple markups in a section', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [
+      ['B'],
+      ['I']
+    ],
     sections: [
-      [        // markers
-        ['B'],
-        ['I']
-      ],
-      [        // sections
-        [MARKUP_SECTION_TYPE, 'P', [
-          [[0], 0, 'hello '], // b
-          [[1], 0, 'brave '], // b+i
-          [[], 1, 'new '], // close i
-          [[], 1, 'world'] // close b
-        ]]
-      ]
+      [MARKUP_SECTION_TYPE, 'P', [
+        [MARKUP_MARKER_TYPE, [0], 0, 'hello '], // b
+        [MARKUP_MARKER_TYPE, [1], 0, 'brave '], // b+i
+        [MARKUP_MARKER_TYPE, [], 1, 'new '], // close i
+        [MARKUP_MARKER_TYPE, [], 1, 'world'] // close b
+      ]]
     ]
   };
   let { result: rendered } = renderer.render(mobiledoc);
   assert.equal(childNodesLength(rendered), 1,
                'renders 1 section');
-  let sectionEl = rendered.childNodes.item(0);
+  let sectionEl = rendered.firstChild;
 
   assert.equal(innerHTML(sectionEl), '<b>hello <i>brave new </i>world</b>');
 });
 
 test('renders a mobiledoc with image section', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [],      // markers
-      [        // sections
-        [IMAGE_SECTION_TYPE, dataUri]
-      ]
+      [IMAGE_SECTION_TYPE, dataUri]
     ]
   };
   let { result: rendered } = renderer.render(mobiledoc);
   assert.equal(childNodesLength(rendered), 1,
                'renders 1 section');
-  let sectionEl = rendered.childNodes.item(0);
+  let sectionEl = rendered.firstChild;
 
   assert.equal(sectionEl.src, dataUri);
 });
 
 test('renders a mobiledoc with built-in image card', (assert) => {
   assert.expect(3);
-  let cardName = 'image';
-  let payload = {
-    src: dataUri
-  };
+  let cardName = ImageCard.name;
+  let payload = { src: dataUri };
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      [cardName, payload]
+    ],
+    markups: [],
     sections: [
-      [],      // markers
-      [        // sections
-        [CARD_SECTION_TYPE, cardName, payload]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
   let { result: rendered } = renderer.render(mobiledoc);
   assert.equal(childNodesLength(rendered), 1,
                'renders 1 section');
-  let sectionEl = rendered.childNodes.item(0);
+  let sectionEl = rendered.firstChild;
 
   assert.equal(sectionEl.tagName, 'IMG');
   assert.equal(sectionEl.src, dataUri);
@@ -199,15 +242,15 @@ test('renders a mobiledoc with built-in image card', (assert) => {
 
 test('render mobiledoc with list section and list items', (assert) => {
   const mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [],
-      [
-        [LIST_SECTION_TYPE, 'ul', [
-          [[[], 0, 'first item']],
-          [[[], 0, 'second item']],
-        ]]
-      ]
+      [LIST_SECTION_TYPE, 'ul', [
+        [[MARKUP_MARKER_TYPE, [], 0, 'first item']],
+        [[MARKUP_MARKER_TYPE, [], 0, 'second item']]
+      ]]
     ]
   };
   let { result: rendered } = renderer.render(mobiledoc, document.createElement('div'));
@@ -218,46 +261,51 @@ test('render mobiledoc with list section and list items', (assert) => {
 
   assert.equal(childNodesLength(section), 2, '2 list items');
 
-  assert.equal(section.firstChild.tagName, 'LI', 'correct tagName for item 1');
-  assert.equal(section.firstChild.firstChild.nodeValue, 'first item',
+  const items = section.childNodes;
+  assert.equal(items.item(0).tagName, 'LI', 'correct tagName for item 1');
+  assert.equal(items.item(0).firstChild.nodeValue, 'first item',
                'correct text node for item 1');
 
-  assert.equal(section.lastChild.tagName, 'LI', 'correct tagName for item 2');
-  assert.equal(section.lastChild.firstChild.nodeValue, 'second item',
+  assert.equal(items.item(1).tagName, 'LI', 'correct tagName for item 2');
+  assert.equal(items.item(1).firstChild.nodeValue, 'second item',
                'correct text node for item 2');
 });
 
 test('renders a mobiledoc with card section', (assert) => {
-  assert.expect(8);
+  assert.expect(9);
   let cardName = 'title-card';
   let expectedPayload = { name: 'bob' };
-  let expectedOptions = {foo: 'bar'};
+  let expectedOptions = { foo: 'bar' };
+  let expectedDom = window.document;
+
   let TitleCard = {
     name: cardName,
     type: 'dom',
-    render({env, options, payload}) {
-      let {name, isInEditor, onTeardown, didRender} = env;
-      assert.equal(name, cardName, 'has name');
-      assert.ok(!isInEditor, 'not isInEditor');
-      assert.ok(!!onTeardown, 'has onTeardown');
-      assert.ok(!!didRender, 'has didRender');
-
-      assert.deepEqual(options, expectedOptions);
-      assert.deepEqual(payload, expectedPayload);
+    render: ({env, payload, options}) => {
+      assert.deepEqual(payload, expectedPayload, 'correct payload');
+      assert.deepEqual(options, expectedOptions, 'correct options');
+      assert.equal(env.name, cardName, 'correct name');
+      assert.ok(!env.isInEditor, 'isInEditor correct');
+      assert.ok(!!env.onTeardown, 'has onTeardown hook');
+      assert.ok(!!env.didRender, 'has didRender hook');
+      assert.deepEqual(env.dom, expectedDom, 'env has dom');
 
       return document.createTextNode(payload.name);
     }
   };
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      [cardName, expectedPayload]
+    ],
+    markups: [],
     sections: [
-      [],      // markers
-      [        // sections
-        [CARD_SECTION_TYPE, cardName, expectedPayload]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
-  renderer = new Renderer({cards: [TitleCard], cardOptions: expectedOptions});
+
+  renderer = new Renderer({cards: [TitleCard], cardOptions: expectedOptions, dom: expectedDom});
   let { result: rendered } = renderer.render(mobiledoc);
   assert.equal(childNodesLength(rendered), 1, 'renders 1 section');
   assert.equal(innerHTML(rendered), expectedPayload.name);
@@ -298,12 +346,14 @@ test('throws if card render returns invalid result', (assert) => {
     render() { return 'string'; }
   };
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      [card.name]
+    ],
+    markups: [],
     sections: [
-      [],
-      [
-        [CARD_SECTION_TYPE, card.name]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
 
@@ -321,12 +371,14 @@ test('card may render nothing', (assert) => {
     render() {}
   };
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      [card.name]
+    ],
+    markups: [],
     sections: [
-      [],
-      [
-        [CARD_SECTION_TYPE, card.name]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
 
@@ -348,24 +400,26 @@ test('rendering nested mobiledocs in cards', (assert) => {
   }];
 
   let innerMobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [], // markers
-      [   // sections
-        [MARKUP_SECTION_TYPE, 'P', [
-          [[], 0, 'hello world']]
-        ]
+      [MARKUP_SECTION_TYPE, 'P', [
+        [MARKUP_MARKER_TYPE, [], 0, 'hello world']]
       ]
     ]
   };
 
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      ['nested-card', {mobiledoc: innerMobiledoc}]
+    ],
+    markups: [],
     sections: [
-      [], // markers
-      [   // sections
-        [CARD_SECTION_TYPE, 'nested-card', {mobiledoc: innerMobiledoc}]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
 
@@ -381,12 +435,14 @@ test('rendering nested mobiledocs in cards', (assert) => {
 test('rendering unknown card without unknownCardHandler throws', (assert) => {
   let cardName = 'not-known';
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      [cardName]
+    ],
+    markups: [],
     sections: [
-      [],
-      [
-        [CARD_SECTION_TYPE, cardName]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
   renderer = new Renderer({cards: [], unknownCardHandler: undefined});
@@ -414,12 +470,14 @@ test('rendering unknown card uses unknownCardHandler', (assert) => {
   };
 
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      [cardName, expectedPayload]
+    ],
+    markups: [],
     sections: [
-      [],
-      [
-        [CARD_SECTION_TYPE, cardName, expectedPayload]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
   renderer = new Renderer({
@@ -451,13 +509,13 @@ test('multiple spaces should preserve whitespace with nbsps', (assert) => {
     repeat(space, 6)
   ].join('');
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [], // markers
-      [   // sections
-        [MARKUP_SECTION_TYPE, 'P', [
-          [[], 0, text]]
-        ]
+      [MARKUP_SECTION_TYPE, 'P', [
+        [MARKUP_MARKER_TYPE, [], 0, text]]
       ]
     ]
   };
@@ -477,9 +535,10 @@ test('multiple spaces should preserve whitespace with nbsps', (assert) => {
 test('throws when given unexpected mobiledoc version', (assert) => {
   let mobiledoc = {
     version: '0.1.0',
-    sections: [
-      [], []
-    ]
+    atoms: [],
+    cards: [],
+    markups: [],
+    sections: []
   };
 
   assert.throws(
@@ -496,17 +555,17 @@ test('throws when given unexpected mobiledoc version', (assert) => {
 
 test('XSS: unexpected markup and list section tag names are not renderered', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [],
-      [
-        [MARKUP_SECTION_TYPE, 'script', [
-          [[], 0, 'alert("markup section XSS")']
-        ]],
-        [LIST_SECTION_TYPE, 'script', [
-          [[[], 0, 'alert("list section XSS")']]
-        ]]
-      ]
+      [MARKUP_SECTION_TYPE, 'script', [
+        [MARKUP_MARKER_TYPE, [], 0, 'alert("markup section XSS")']
+      ]],
+      [LIST_SECTION_TYPE, 'script', [
+        [[MARKUP_MARKER_TYPE, [], 0, 'alert("list section XSS")']]
+      ]]
     ]
   };
   let { result } = renderer.render(mobiledoc);
@@ -516,20 +575,20 @@ test('XSS: unexpected markup and list section tag names are not renderered', (as
 
 test('XSS: unexpected markup types are not rendered', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [
+      ['b'], // valid
+      ['em'], // valid
+      ['script'] // invalid
+    ],
     sections: [
-      [
-        ['b'], // valid
-        ['em'], // valid
-        ['script'] // invalid
-      ],
-      [
-        [MARKUP_SECTION_TYPE, 'p', [
-          [[0], 0, 'bold text'],
-          [[1,2], 3, 'alert("markup XSS")'],
-          [[], 0, 'plain text']
-        ]]
-      ]
+      [MARKUP_SECTION_TYPE, 'p', [
+        [MARKUP_MARKER_TYPE, [0], 0, 'bold text'],
+        [MARKUP_MARKER_TYPE, [1,2], 3, 'alert("markup XSS")'],
+        [MARKUP_MARKER_TYPE, [], 0, 'plain text']
+      ]]
     ]
   };
   let { result } = renderer.render(mobiledoc);
@@ -540,17 +599,22 @@ test('XSS: unexpected markup types are not rendered', (assert) => {
 test('XSS: links with dangerous href values are not rendered', (assert) => {
   let unsafeHref = 'javascript:alert("link XSS")'; // jshint ignore:line
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
-    sections: [
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [
       [
-        ["a", [ "href", unsafeHref ]]
-      ],
-      [
-        [MARKUP_SECTION_TYPE, "p", [
-          [[0], 1, "link text"],
-          [[], 0, "plain text"]]
+        'a', [
+          'href',
+          unsafeHref
         ]
       ]
+    ],
+    sections: [
+      [MARKUP_SECTION_TYPE, 'p', [
+        [MARKUP_MARKER_TYPE, [0], 1, 'link text'],
+        [MARKUP_MARKER_TYPE, [], 0, 'plain text']
+      ]]
     ]
   };
   let { result } = renderer.render(mobiledoc);
@@ -558,21 +622,192 @@ test('XSS: links with dangerous href values are not rendered', (assert) => {
   assert.equal(content, `<p><a href="unsafe:${escapeQuotes(unsafeHref)}">link text</a>plain text</p>`);
 });
 
+test('renders a mobiledoc with atom', (assert) => {
+  assert.expect(8);
+  let atomName = 'hello-atom';
+  let expectedPayload = { name: 'bob' };
+  let expectedOptions = { foo: 'bar' };
+  let expectedValue = '@BOB';
+  let expectedDom = window.document;
+
+  let atom = {
+    name: atomName,
+    type: 'dom',
+    render({ env, payload, value, options }) {
+      assert.deepEqual(payload, expectedPayload, 'correct payload');
+      assert.deepEqual(options, expectedOptions, 'correct options');
+      assert.equal(value, expectedValue, 'correct value');
+      assert.equal(env.name, atomName, 'correct name');
+      assert.ok(!env.isInEditor, 'isInEditor correct');
+      assert.ok(!!env.onTeardown, 'has onTeardown hook');
+      assert.deepEqual(env.dom, expectedDom, 'env has dom');
+
+      return document.createTextNode(`Hello ${value}`);
+    }
+  };
+  let mobiledoc = {
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [
+      ['hello-atom', expectedValue, expectedPayload],
+    ],
+    cards: [],
+    markups: [],
+    sections: [
+      [MARKUP_SECTION_TYPE, 'P', [
+        [ATOM_MARKER_TYPE, [], 0, 0]]
+      ]
+    ]
+  };
+  renderer = new Renderer({atoms: [atom], cardOptions: expectedOptions, dom: expectedDom});
+  let { result: rendered } = renderer.render(mobiledoc);
+
+  let sectionEl = rendered.firstChild;
+  assert.equal(sectionEl.textContent, 'Hello ' + expectedValue);
+});
+
+
+test('throws when given atom with invalid type', (assert) => {
+  let atom = {
+    name: 'bad',
+    type: 'other',
+    render() {}
+  };
+  assert.throws(
+    () => { new Renderer({atoms: [atom]}); }, // jshint ignore:line
+    /Atom "bad" must be type "dom"/
+  );
+});
+
+test('throws when given atom without `render`', (assert) => {
+  let atom = {
+    name: 'bad',
+    type: 'dom',
+    render: undefined
+  };
+  assert.throws(
+    () => { new Renderer({atoms: [atom]}); }, // jshint ignore:line
+    /Atom "bad" must define.*render/
+  );
+});
+
+test('throws if atom render returns invalid result', (assert) => {
+  let atom = {
+    name: 'bad',
+    type: 'dom',
+    render() { return 'string'; }
+  };
+  let mobiledoc = {
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [
+      ['bad', 'Bob', { id: 42 }],
+    ],
+    cards: [],
+    markups: [],
+    sections: [
+      [MARKUP_SECTION_TYPE, 'P', [
+        [ATOM_MARKER_TYPE, [], 0, 0]]
+      ]
+    ]
+  };
+  renderer = new Renderer({atoms: [atom]});
+  assert.throws(
+    () => renderer.render(mobiledoc),
+    /Atom "bad" must render dom/
+  );
+});
+
+test('atom may render nothing', (assert) => {
+  let atom = {
+    name: 'ok',
+    type: 'dom',
+    render() {}
+  };
+  let mobiledoc = {
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [
+      ['ok', 'Bob', { id: 42 }],
+    ],
+    cards: [],
+    markups: [],
+    sections: [
+      [MARKUP_SECTION_TYPE, 'P', [
+        [ATOM_MARKER_TYPE, [], 0, 0]]
+      ]
+    ]
+  };
+
+  renderer = new Renderer({atoms:[atom]});
+  renderer.render(mobiledoc);
+
+  assert.ok(true, 'No error thrown');
+});
+
+test('throws when rendering unknown atom without unknownAtomHandler', (assert) => {
+  let mobiledoc = {
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [
+      ['missing-atom', 'Bob', { id: 42 }],
+    ],
+    cards: [],
+    markups: [],
+    sections: [
+      [MARKUP_SECTION_TYPE, 'P', [
+        [ATOM_MARKER_TYPE, [], 0, 0]]
+      ]
+    ]
+  };
+  renderer = new Renderer({atoms: [], unknownAtomHandler: undefined});
+  assert.throws(
+    () => renderer.render(mobiledoc),
+    /Atom "missing-atom" not found.*no unknownAtomHandler/
+  );
+});
+
+test('rendering unknown atom uses unknownAtomHandler', (assert) => {
+  assert.expect(4);
+
+  let atomName = 'missing-atom';
+  let expectedPayload = { id: 42 };
+  let cardOptions = {};
+  let mobiledoc = {
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [
+      ['missing-atom', 'Bob', { id: 42 }],
+    ],
+    cards: [],
+    markups: [],
+    sections: [
+      [MARKUP_SECTION_TYPE, 'P', [
+        [ATOM_MARKER_TYPE, [], 0, 0]]
+      ]
+    ]
+  };
+  let unknownAtomHandler = ({env, payload, options}) => {
+    assert.equal(env.name, atomName, 'correct name');
+    assert.ok(!!env.onTeardown, 'onTeardown hook exists');
+
+    assert.deepEqual(payload, expectedPayload, 'correct payload');
+    assert.deepEqual(options, cardOptions, 'correct options');
+  };
+  renderer = new Renderer({atoms: [], unknownAtomHandler, cardOptions});
+  renderer.render(mobiledoc);
+});
+
 test('renders a mobiledoc with sectionElementRenderer', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [], // markers
-      [   // sections
       [MARKUP_SECTION_TYPE, 'P', [
-        [[], 0, 'hello world']]
+        [MARKUP_MARKER_TYPE, [], 0, 'hello world']]
       ],
       [MARKUP_SECTION_TYPE, 'p', [
-        [[], 0, 'hello world']]
+        [MARKUP_MARKER_TYPE, [], 0, 'hello world']]
       ],
       [MARKUP_SECTION_TYPE, 'h1', [
-        [[], 0, 'hello world']]
-      ]
+        [MARKUP_MARKER_TYPE, [], 0, 'hello world']]
       ]
     ]
   };
@@ -592,23 +827,23 @@ test('renders a mobiledoc with sectionElementRenderer', (assert) => {
                'renders the text');
   assert.equal(rendered.childNodes.item(1).tagName, 'PRE',
                'renders a pre');
-  assert.equal(rendered.lastChild.tagName, 'H2',
+  assert.equal(rendered.childNodes.item(2).tagName, 'H2',
                'renders an h2');
 });
 
 test('renders a mobiledoc with markupElementRenderer', (assert) => {
   let mobiledoc = {
-    "version": MOBILEDOC_VERSION,
+    "version": MOBILEDOC_VERSION_0_3_0,
+    "atoms": [],
+    "cards": [],
+    "markups": [
+      ["a", [ "href", "#foo" ]]
+    ],
     "sections": [
-      [
-        ["A", [ "href", "#foo" ]]
-      ],
-      [
-        [MARKUP_SECTION_TYPE, "p", [
-          [[], 0, "Lorem ipsum "],
-          [[0], 1, "dolor"],
-          [[], 0, " sit amet."]]
-        ]
+      [MARKUP_SECTION_TYPE, "p", [
+        [MARKUP_MARKER_TYPE, [], 0, "Lorem ipsum "],
+        [MARKUP_MARKER_TYPE, [0], 1, "dolor"],
+        [MARKUP_MARKER_TYPE, [], 0, " sit amet."]]
       ]
     ]
   };
@@ -629,7 +864,7 @@ test('renders a mobiledoc with markupElementRenderer', (assert) => {
                'renders text inside of marker');
   assert.equal(rendered.firstChild.childNodes[1].tagName, 'SPAN',
                'transforms markup nodes');
-  assert.propEqual(rendered.firstChild.childNodes[1].dataset, {tag: "A", href: "#foo"},
+  assert.propEqual(rendered.firstChild.childNodes[1].dataset, {tag: "a", href: "#foo"},
                    'passes original tag and attributes to transform');
   assert.equal(rendered.firstChild.childNodes[0].textContent, 'Lorem ipsum ',
                'renders plain text nodes');
@@ -637,23 +872,23 @@ test('renders a mobiledoc with markupElementRenderer', (assert) => {
                'renders text nodes as proper type');
 });
 
-test('unexpected markup types are not passed to markup renderer', (assert) => {
+test('unexpected markup types are not handled by markup renderer', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [
+      ['script']
+    ],
     sections: [
-      [
-        ['script'] // invalid
-      ],
-      [
-        [MARKUP_SECTION_TYPE, 'p', [
-          [[0], 1, 'alert("markup XSS")']
-        ]]
-      ]
+      [MARKUP_SECTION_TYPE, 'p', [
+        [MARKUP_MARKER_TYPE, [0], 1, 'alert("markup XSS")']
+      ]]
     ]
   };
   renderer = new Renderer({
     markupElementRenderer: {
-      SCRIPT: (tagName, dom) => {
+      SCRIPT: (markerType, dom) => {
         return dom.createElement('script');
       }
     }
@@ -664,7 +899,7 @@ test('unexpected markup types are not passed to markup renderer', (assert) => {
 });
 }
 
-module('Unit: Mobiledoc DOM Renderer - 0.2', {
+module('Unit: Mobiledoc DOM Renderer - 0.3', {
   beforeEach() {
     renderer = new Renderer();
   }
@@ -674,14 +909,14 @@ generateTests();
 
 test('teardown removes rendered sections from dom', (assert) => {
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [],
+    markups: [],
     sections: [
-      [],
-      [
-        [MARKUP_SECTION_TYPE, 'p', [
-          [[], 0, 'Hello world']
-        ]]
-      ]
+      [MARKUP_SECTION_TYPE, 'p', [
+        [MARKUP_MARKER_TYPE, [], 0, 'Hello world']
+      ]]
     ]
   };
 
@@ -711,12 +946,14 @@ test('teardown hook calls registered teardown methods', (assert) => {
   };
 
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      [cardName]
+    ],
+    markups: [],
     sections: [
-      [],
-      [
-        [CARD_SECTION_TYPE, cardName]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
 
@@ -743,12 +980,14 @@ test('render hook calls registered didRender callbacks', (assert) => {
   };
 
   let mobiledoc = {
-    version: MOBILEDOC_VERSION,
+    version: MOBILEDOC_VERSION_0_3_0,
+    atoms: [],
+    cards: [
+      [cardName]
+    ],
+    markups: [],
     sections: [
-      [],
-      [
-        [CARD_SECTION_TYPE, cardName]
-      ]
+      [CARD_SECTION_TYPE, 0]
     ]
   };
 
@@ -758,11 +997,10 @@ test('render hook calls registered didRender callbacks', (assert) => {
 
   renderer.render(mobiledoc);
 
-  assert.ok(didRender, 'didRender called');
+  assert.ok(!!didRender, 'didRender called');
 });
 
-
-module('Unit: Mobiledoc DOM Renderer w/ SimpleDOM - 0.2', {
+module('Unit: Mobiledoc DOM Renderer w/ SimpleDOM - 0.3', {
   beforeEach() {
     renderer = new Renderer({ dom: new SimpleDOM.Document() });
   }
