--- tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "target": "es2017",
-    "module": "es2015",
-    "moduleResolution": "node",
-    "experimentalDecorators": true,
-    "inlineSources": true,
-    "inlineSourceMap": true,
-    "newLine": "LF",
-    "outDir": "dist"
-  },
-  "include": [
-    "lib/**/*"
-  ]
-}
