--- tsconfig.tests.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "target": "es5",
-    "module": "amd",
-    "experimentalDecorators": true,
-    "moduleResolution": "node",
-    "inlineSources": true,
-    "inlineSourceMap": true,
-    "declaration": false,
-    "newLine": "LF",
-    "outFile": "tests/tests.js",
-    "types": ["qunit"]
-  },
-  "include": [
-    "lib/**/*",
-    "test/**/*",
-    "node_modules/@types/**/*.ts"
-  ]
-}
